add animation end force flag
*) sometimes,we wonna to know is the animation end by client or itself.
for this, i add a boolean value to provide the msg.
*) change some class name to make the code easy read.
*) move the myapplication.class to package "application",instead of the
package "common"
*) add skill type to idenfity a skill is active_attack or buff or debuff
or passive. it would useful in some time.
*) add propetry observer into package "observer",we will need notify the
observer in some time,eg:client will notify the propetry view and blood
view when the player be injured.
*) remove the statement "if(propetry.getMagicVolume() > 0)". For drawing
the magic and blood reservoir,it needn't the statement anymore.