optimize the map and monster
map can be indicated in external xml.
monster have the same feature too.
abstract the device screem to camera, and the  camera have width, when
move the player let the camera follow the player so that there would
have a pefert animation.
when the camera touch the boundary,think the camera's actual width,the
player will move itself to the boundary.